requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
cffi==1.16.0
channels==3.0.5
channels-redis==4.2.0
constantly==23.10.4
cryptography==42.0.5
daphne==3.0.2
Django==4.2.11
django-cors-headers==4.3.1
django-redis==5.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
msgpack==1.0.8
psycopg2==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==24.1.0
pytz==2024.1
redis==5.0.2
service-identity==24.1.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
zope.interface==6.2
